Rethink root collection build.
