+++ services/canopy.js
@@ -2,7 +2,202 @@
  *
  */
 
+const slugify = require("slugify");
+
+const slugifyConfig = {
+  lower: true,
+  strict: true,
+  trim: true,
+};
+
 module.exports.buildCanopy = (env) => {
+  console.log();
   console.log(`env`, env);
-  console.log(`build`);
+
+  const rootCollection = getCollection(0, env.collection);
+
+  // const collection = iiif.getCollectionData();
+  console.log(`rootCollection`, rootCollection);
+  // console.log(`build`);
+};
+
+/**
+ *
+ * @param json
+ * @param depth
+ * @param parent
+ * @returns
+ */
+const buildCollection = (json, depth, parent = null) => {
+  if (!json) return null;
+
+  /**
+   * defensively determine @context
+   */
+  let context = json["@context"];
+  if (!context) context = "http://iiif.io/api/presentation/3/context.json";
+
+  if (Array.isArray(context)) context = context[0];
+  context = context.replace("https://", "");
+  context = context.replace("http://", "");
+
+  /**
+   * set props
+   */
+  let id = null;
+  let label = null;
+  let slug = null;
+  let children = null;
+
+  /**
+   * based on @context, parse collections
+   */
+  switch (context) {
+    case "iiif.io/api/presentation/2/context.json":
+      id = json["@id"];
+      label = getLabel(json.label);
+      slug = slugify(label[0], { ...slugifyConfig });
+      children = buildCollectionItems2(json, id);
+      return {
+        id,
+        label,
+        slug,
+        depth,
+        parent,
+        manifests: children.manifests,
+        collections: children.collections,
+        items: children.items,
+      };
+    case "iiif.io/api/presentation/3/context.json":
+      id = json.id;
+      label = getLabel(json.label);
+      slug = slugify(label[0], { ...slugifyConfig });
+      children = buildCollectionItems3(json, id);
+      return {
+        id,
+        label,
+        slug,
+        depth,
+        parent,
+        manifests: children.manifests,
+        collections: children.collections,
+        items: children.items,
+      };
+  }
+
+  console.log(children);
+
+  return {
+    id,
+    label,
+    slug,
+    depth,
+    parent,
+    manifests: children.manifests,
+    collections: children.collections,
+    items: children.items,
+  };
+};
+
+/**
+ *
+ * @param items
+ * @param parent
+ * @returns
+ */
+const buildCollectionItems2 = (json, parent) => {
+  let manifests = 0;
+  let collections = 0;
+  let items = [];
+
+  if (json.collections)
+    items = items.concat(
+      json.collections.map((item) => {
+        item.id = item["@id"];
+        delete item["@id"];
+
+        item.type = item["@type"].replace("sc:", "");
+        delete item["@type"];
+
+        if (item.type === "Manifest") manifests++;
+        if (item.type === "Collection") collections++;
+        item.label = getLabel(item.label);
+        item.parent = parent;
+        return item;
+      })
+    );
+
+  if (json.manifests)
+    items = items.concat(
+      json.manifests.map((item) => {
+        item.id = item["@id"];
+        delete item["@id"];
+
+        item.type = item["@type"].replace("sc:", "");
+        delete item["@type"];
+
+        if (item.type === "Manifest") manifests++;
+        if (item.type === "Collection") collections++;
+        item.label = getLabel(item.label);
+        item.parent = parent;
+        return item;
+      })
+    );
+
+  return {
+    items: items,
+    manifests,
+    collections,
+  };
+};
+
+const buildCollectionItems3 = (json, parent) => {
+  let manifests = 0;
+  let collections = 0;
+  return {
+    items: json.items.map((item) => {
+      if (item.type === "Manifest") manifests++;
+      if (item.type === "Collection") collections++;
+      item.label = getLabel(item.label);
+      item.parent = parent;
+      return item;
+    }),
+    manifests,
+    collections,
+  };
+};
+
+const getCollection = (depth, id, parent = null) =>
+  fetch(id)
+    .then((response) => response.text())
+    .then((text) => {
+      try {
+        return JSON.parse(text);
+      } catch (err) {
+        console.log(err);
+      }
+    })
+    .then((json) => buildCollection(json, depth, parent));
+
+const getLabel = (label, language = "none") => {
+  /*
+   * If no label exists, return a hardcoded string.
+   */
+  if (!label) return ["Untitled"];
+
+  if (typeof label === "string") return [label];
+
+  /*
+   * If InternationalString code does not exist on label, then
+   * return what may be there, ex: label.none[0] OR label.fr[0]
+   */
+  if (!label[language]) {
+    const codes = Object.getOwnPropertyNames(label);
+    if (codes.length > 0) return label[codes[0]];
+  }
+
+  /*
+   * Return label value for InternationalString code `en`
+   */
+  return label[language];
 };
